+++ frog/imports/ui/GraphEditor/SidePanel/FormWidgets/index.js
@@ -1,0 +1,5 @@
+// @flow
+
+export { default as SelectActivityWidget } from './SelectActivity';
+export { default as SelectAttributeWidget } from './SelectAttribute';
+export { default as SelectFormWidget } from './SelectForm';
